background remover as a library
