+++ .gitignore
@@ -1,3 +1,4 @@
+.idea/
 # Created by https://www.toptal.com/developers/gitignore/api/python
 # Edit at https://www.toptal.com/developers/gitignore?templates=python
