+++ README.md
@@ -137,6 +137,27 @@ change the model for different background removal methods between `u2netp`, `u2n
 backgroundremover -i "/path/to/video.mp4" -m "u2net_human_seg" -fl 150 -tv -o "output.mov"
 ```
 
+## As a library
+### Remove background image
+
+```
+from backgroundremover.bg import remove
+def remove_bg(src_img_path, out_img_path):
+    model_choices = ["u2net", "u2net_human_seg", "u2netp"]
+    f = open(src_img_path, "rb")
+    data = f.read()
+    img = remove(data, model_name=model_choices[0],
+                 alpha_matting=True,
+                 alpha_matting_foreground_threshold=240,
+                 alpha_matting_background_threshold=10,
+                 alpha_matting_erode_structure_size=10,
+                 alpha_matting_base_size=1000)
+    f.close()
+    f = open(out_img_path, "wb")
+    f.write(img)
+    f.close()
+```
+
 ## Todo
 
 - convert logic from video to image to utilize more GPU on image removal
